+++ src/Router.jsx
@@ -10,7 +10,7 @@ import PostPage from "./pages/PostPage";
 
 import Header from "./layouts/Header/Header";
 import MainLayout from "./layouts/MainLayout/MainLayout";
-import PostItemPageHeader from "./layouts/PostItemPageHeader PostItemPageHeader PostItemPageHeader/PostItemPageHeader";
+import PostItemPageHeader from "./layouts/PostItemPageHeader/PostItemPageHeader";
 
 import { ROUTES } from "./constants/routes";
